requirements.txt
torch>=2.10.0
torchvision>=0.25.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
Pillow>=9.0.0
tqdm>=4.64.0
